+++ src/pages/Newsletter/containers/NewsletterContainer.tsx
@@ -8,7 +8,7 @@ type Props = {};
 
 const NewsletterContainer = (props: Props) => {
 	const [newsletterList, setNewsletterList] = useState<Newsletters[]>([]);
-	const downloadExcel = (data: Newsletters[]) => {
+	const downloadExcel = useCallback((data: Newsletters[]) => {
 		// 데이터를 이차원 배열로 변환
 		const dataArray = [
 			['Index', 'ID', 'Email'], // 라벨 배열
@@ -27,16 +27,21 @@ const NewsletterContainer = (props: Props) => {
 
 		// 엑셀 파일 생성 및 다운로드
 		XLSX.writeFile(wb, 'newsletter.xlsx');
-	};
+	}, []);
+
 	const onTrashClick = useCallback(
 		(id: number) => {
-			setNewsletterList((prevList: Newsletters[]) => {
-				const updatedList = prevList.filter(
-					(newsletter) => newsletter.id !== id,
-				);
-				return updatedList;
-			});
-			deleteNewsletter(id);
+			const confirmed = window.confirm('삭제하시겠습니까?');
+			if (confirmed) {
+				setNewsletterList((prevList: Newsletters[]) => {
+					const updatedList = prevList.filter(
+						(newsletter) => newsletter.id !== id,
+					);
+					return updatedList;
+				});
+				deleteNewsletter(id);
+				alert('삭제되었습니다');
+			}
 		},
 		[newsletterList],
 	);
@@ -49,18 +54,6 @@ const NewsletterContainer = (props: Props) => {
 
 		return () => {};
 	}, []);
-	// useEffect(() => {
-	// 	const updatedList = [];
-	// 	for (let i = 0; i < 8; i++) {
-	// 		const newsletter = {
-	// 			id: i + 1,
-	// 			email: 'fjal;k@asfmdal.o.kr',
-	// 		};
-	// 		updatedList.push(newsletter);
-	// 	}
-	// 	setNewsletterList(updatedList);
-	// 	return () => {};
-	// }, []);
 
 	return (
 		<Newsletter
